refactor: utils folder to organize classes
